requirements.txt
colorama==0.4.6
filelock==3.16.0
fsspec==2024.9.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.1
pandas==2.2.2
pillow==10.4.0
python-dateutil==2.9.0.post0
pytz==2024.2
setuptools==75.1.0
six==1.16.0
sympy==1.13.2
torch==2.4.1+cu118
torchaudio==2.4.1+cu118
torchvision==0.19.1+cu118
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
